requirements.txt
anyio==3.7.1
asgiref==3.8.1
atomicwrites==1.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.34.136
botocore==1.34.136
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
constantly==23.10.4
coverage==7.5.4
cryptography==42.0.8
daphne==4.1.2
dataclasses==0.6
defusedxml==0.8.0rc2
deprecation==2.1.0
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-filter==24.2
django-heroku==0.3.1
django-oauth-toolkit==2.4.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.2.0
django-rest-passwordreset==1.4.1
django-storage-supabase==0.1.0
django-storages==1.14.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
flake8==7.1.0
gotrue==0.2.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jwcrypto==1.5.6
mccabe==0.7.0
mock==5.1.0
msgpack==1.0.8
nose==1.3.7
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
pipdeptree==2.23.0
pluggy==1.5.0
postgrest==0.10.6
postgrest-py==0.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==1.10.17
pyflakes==3.2.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytest==6.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
realtime==1.0.6
realtime-py==0.1.3
redis==5.0.7
requests==2.25.1
requests-oauthlib==2.0.0
rfc3986==1.5.0
s3transfer==0.10.2
service-identity==24.1.0
setuptools==70.1.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
storage3==0.7.6
StrEnum==0.4.15
supabase==0.0.3
supafunc==0.4.6
toml==0.10.2
tqdm==4.66.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
uvicorn==0.30.1
websockets==9.1
whitenoise==6.7.0
zope.interface==6.4.post2
